Refactor of package constraints
