+++ src/main/java/org/jonleb/characters/constraints/diccetower/DiceTowerValidator.java
@@ -1,4 +1,4 @@
-package org.jonleb.characters.constraints;
+package org.jonleb.characters.constraints.diccetower;
 
 import lombok.extern.log4j.Log4j2;
