+++ src/main/java/org/jonleb/characters/services/DiceTower.java
@@ -1,8 +1,8 @@
 package org.jonleb.characters.services;
 
 import lombok.extern.log4j.Log4j2;
-import org.jonleb.characters.constraints.DiceTowerConstraint;
-import org.jonleb.characters.constraints.RollConstraint;
+import org.jonleb.characters.constraints.diccetower.DiceTowerConstraint;
+import org.jonleb.characters.constraints.diccetower.RollConstraint;
 import org.jonleb.characters.utils.DiceRoller;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
@@ -11,6 +11,7 @@ import javax.validation.Valid;
 import java.util.HashMap;
 import java.util.Map;
 
+@SuppressWarnings("DuplicatedCode")
 @Log4j2
 @Service
 @Validated
